+++ backend/src/main/java/com/agrupae/application/exception/assignment/NotCourseLeaderException.java
@@ -1,0 +1,9 @@
+package com.agrupae.application.exception.assignment;
+
+import com.agrupae.application.exception.ApplicationException;
+
+public class NotCourseLeaderException extends ApplicationException {
+    public NotCourseLeaderException(String message) {
+        super(message);
+    }
+}